Maintain local copies of the current depth
ActiveSupport will call `merge` on the GeneratorState, which resets
the depth back to 0. Then we decrement it again and pass that negative
value to System.arraycopy, which throws an exception.

This fixes the problem in the same way as the C extension: it
maintains local copies of the depth value, and resets the state back
when it is done with each nested object.