Never allow the current depth to be negative
ActiveSupport will call `merge` on the GeneratorState, which resets
the depth back to 0. Then we decrement it again and pass that negative
value to System.arraycopy, which throws an exception.

This fixes the problem by ignoring it; if the depth ever goes
negative, we pretend it is zero and stop indenting. This results in
less-than-pretty output, but is very easy to reason about.